agent_system/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
PyYAML==6.0.2
pytest==8.4.1
httpx==0.28.1
pypdf==5.4.0
